+++ packages/upnp-nat/src/static-gateway-finder.ts
@@ -1,5 +1,5 @@
 import { TypedEventEmitter } from 'main-event'
-import type { GatewayFinder, GatewayFinderEvents } from './upnp-nat.js'
+import type { GatewayFinder, GatewayFinderEvents } from './upnp-nat.ts'
 import type { Gateway, UPnPNAT } from '@achingbrain/nat-port-mapper'
 import type { ComponentLogger, Logger } from '@libp2p/interface'